resolve logic conflict

+++ Assets/Src/Framework/TimeDefine.cs
@@ -7,5 +7,5 @@ public static class TimeDefine
     /// <summary>
     /// 毫秒转秒
     /// </summary>
-    public const int MS_2_S = 1 / S_2_MS;
+    public const float MS_2_S = 1.0f / S_2_MS;
 }
